+++ tests/_tplcache/demo_tpl.py
@@ -47,7 +47,7 @@ def __getTplClass__(Contemplate):
                     __instance__.data['i'] = i
                     __instance__.data['usergroup'] = usergroup
                      
-                    __p__ += '' + "\n" + '        <!-- call a (sub-)template -->' + "\n" + '        ' + Contemplate.tpl( "sub",  {"i" : __instance__.data['i'], "users" : __instance__.data['users']} ) 
+                    __p__ += '' + "\n" + '        ' + str( Contemplate.count(__instance__.data['users'][__instance__.data['i']]) ) + "\n" + '        <!-- call a (sub-)template -->' + "\n" + '        ' + Contemplate.tpl( "sub",  {"i" : __instance__.data['i'], "users" : __instance__.data['users']} ) 
                      
                     __p__ += '' + "\n" + '    ' 
              
@@ -61,18 +61,10 @@ def __getTplClass__(Contemplate):
             
             __p__ = ''
              
-            __p__ += '' + "\n" + "\n" + '    <strong>Block2 is overriden by the demo template</strong>' + "\n" + "\n" + '    <br /><br />' + "\n" + "\n" + '    <strong>A table</strong><br />' + "\n" + '    ' + Contemplate.htmltable(__instance__.data['table_data'], __instance__.data['table_options']) 
-             
-            __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>Set a new tpl variable and use it in a custom plugin</strong><br />' + "\n" + '    ';
+            __p__ += '' + "\n" + "\n" + '    <strong>Block2 is overriden by the demo template</strong>' + "\n" + "\n" + '    <br /><br />' + "\n" + "\n" + '    <strong>A table</strong><br />' + "\n" + '    ' + str( Contemplate.htmltable(__instance__.data['table_data'], __instance__.data['table_options']) ) + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>Set a new tpl variable and use it in a custom test plugin</strong><br />' + "\n" + '    ';
             __instance__.data['foo'] = ("123")
              
-            __p__ += '' + "\n" + '    ' + "\n" + '    ' + str( Contemplate.plugin_test(__instance__.data['foo']) ) + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>A select box</strong><br />' + "\n" + '    ' + Contemplate.htmlselect(__instance__.data['select_data'], __instance__.data['select_options']) 
-             
-            __p__ += '' + "\n" + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>A table with alternative format</strong><br />' + "\n" + '    ' + Contemplate.htmltable(__instance__.data['table_data'], {"header" : True}) 
-             
-            __p__ += '' + "\n" + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>A select box with alternative format</strong><br />' + "\n" + '    ' + Contemplate.htmlselect(__instance__.data['select_data'], {             "optgroups" : ["group1", "group2", "group3"],             "selected" : 3,             "multiple" : False,             "style" : "width:200px;",             "foo123" : ":,=>"         }) 
-             
-            __p__ += '' + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    <!-- include a (sub-)template file -->' + "\n" + '     <!-- print a localized date php-style -->' + "\n" + '<strong>A (localized) date, PHP-style</strong><br />' + "\n" + str( Contemplate.ldate("M, d", Contemplate.now()) ) + "\n" + '' 
+            __p__ += '' + "\n" + '    ' + "\n" + '    ' + str( Contemplate.plugin_test(__instance__.data['foo']) ) + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <!-- use contemplate literal data with template function, in this case a print test plugin -->' + "\n" + '    <strong>use literal data with template function, in this case a print plugin</strong><br />' + "\n" + '    ' + str( Contemplate.plugin_print({          "stringVar"     : "stringValue",          "numericVar"    : 123,          "arrayVar"      : [             0, 1, "astring", 3,              { "prop": 1 }          ]      }) ) + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>A select box</strong><br />' + "\n" + '    ' + str( Contemplate.htmlselect(__instance__.data['select_data'], __instance__.data['select_options']) ) + "\n" + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>A table with alternative format</strong><br />' + "\n" + '    ' + str( Contemplate.htmltable(__instance__.data['table_data'], {"header" : True}) ) + "\n" + "\n" + '    <br /><br />' + "\n" + '    ' + "\n" + '    <strong>A select box with alternative format</strong><br />' + "\n" + '    ' + str( Contemplate.htmlselect(__instance__.data['select_data'], {             "optgroups" : ["group1", "group2", "group3"],             "selected" : 3,             "multiple" : False,             "style" : "width:200px;",             "foo123" : ":,=>"         }) ) + "\n" + '    ' + "\n" + '    <br /><br />' + "\n" + '    <!-- include a (sub-)template file -->' + "\n" + '     <!-- print a localized date php-style -->' + "\n" + '<strong>A (localized) date, PHP-style</strong><br />' + "\n" + str( Contemplate.ldate("M, d", Contemplate.now()) ) + "\n" + '' 
              
             __p__ += '' + "\n" + "\n" + ''
             return __p__